requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.30
botocore==1.19.30
CacheControl==0.12.6
cachetools==4.1.1
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.0.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.1.4
django-cors-headers==3.5.0
django-decouple==2.1
django-filter==2.4.0
django-jazzmin==2.4.1
django-rest-swagger==2.2.0
django-storages==1.10.1
django-summernote==0.8.11.6
djangorestframework==3.12.2
drf-yasg==1.17.1
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.0
openapi-codec==1.3.2
packaging==20.9
Pillow==8.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
python-dateutil==2.8.1
python-http-client==3.3.1
python-magic==0.4.22
pytz==2020.4
requests==2.25.0
rsa==4.6
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sendgrid==6.4.8
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tinycss2==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
zipp==3.4.0
